docs: added a lot of comments on the further course of action
It seems, that the current code is very temporary. Now that I know how to convert html to mp4 with Puppeteer, I need html instead of txt. It shouldn't be a problem with current code.
If crawling with Puppeteer really would be more reliable, that's what I'll do. I can scrape YT for the ammount of views on the first couple of videos from a prompt which is an SO title.