fix: to get actual recent file number
the main loop's design was pretty strange, as well as it was decreasing with time. Fixing inconcistency with it and a funcion in src solved the issue. A quarter of pts used!
Next probably video from it all. Have to add some style it all that first, also add N of ppl it helped in there somewhere.

As it is, I think that the crawling part is done, and rn I'll start transforming it into mp4. Fixes #33